+++ src/fmc.rs
@@ -6,9 +6,9 @@
 use stm32_fmc::FmcPeripheral;
 use stm32_fmc::{AddressPinSet, PinsSdram, Sdram, SdramChip, SdramPinSet, SdramTargetBank};
 
-use crate::embedded_time::rate::Hertz;
 use crate::pac;
 use crate::rcc::{Clocks, Enable, Reset};
+use fugit::HertzU32 as Hertz;
 
 use crate::gpio::{self, Alternate};
 
@@ -75,7 +75,7 @@ unsafe impl FmcPeripheral for FMC {
 
     fn source_clock_hz(&self) -> u32 {
         // FMC block is clocked by HCLK
-        self.hclk.0
+        self.hclk.raw()
     }
 }